backend/dms_backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
google-cloud-vision==3.5.0
google-cloud-language==2.13.0
python-dotenv==1.0.0
pydantic==2.5.1
pydantic-settings==2.1.0
bcrypt==4.0.1
requests==2.31.0
aiohttp==3.9.1
numpy==1.26.2
scikit-learn==1.3.2
transformers==4.35.2
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.0+cpu
nltk==3.8.1
spacy==3.7.2
